Add multiplatform support to coil zoomimage
