+++ zoomimage-compose-coil-core/build.gradle.kts
@@ -1,6 +1,42 @@
 plugins {
+    alias(libs.plugins.org.jetbrains.kotlin.multiplatform)
     alias(libs.plugins.com.android.library)
-    alias(libs.plugins.org.jetbrains.kotlin.android)
+    alias(libs.plugins.org.jetbrains.compose)
+}
+
+kotlin {
+    androidTarget {
+        compilations.configureEach {
+            kotlinOptions {
+                jvmTarget = "1.8"
+            }
+        }
+    }
+
+    jvm("desktop") {
+        compilations.configureEach {
+            kotlinOptions {
+                jvmTarget = "1.8"
+            }
+        }
+    }
+
+    sourceSets {
+        named("commonMain") {
+            dependencies {
+                api(project(":zoomimage-compose"))
+                api(project(":zoomimage-core-coil"))
+                api(libs.coil.compose.core)
+            }
+        }
+        named("commonTest") {
+            dependencies {
+                implementation(kotlin("test"))
+                implementation(libs.junit)
+                implementation(libs.panpf.tools4j.test)
+            }
+        }
+    }
 }
 
 android {
@@ -15,23 +51,19 @@ android {
     buildFeatures {
         compose = true
     }
-    composeOptions {
-        kotlinCompilerExtensionVersion = libs.versions.androidx.compose.compiler.get()
-    }
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
     }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
+}
+
+compose {
+    val compilerDependencyDeclaration =
+        libs.androidx.compose.compiler.get().run { "$module:$version" }
+    kotlinCompilerPlugin.set(compilerDependencyDeclaration)
 }
 
 dependencies {
-    api(project(":zoomimage-compose"))
-    api(project(":zoomimage-core-coil"))
-    api(libs.coil.compose.core)
-
     debugImplementation(libs.androidx.compose.ui.tooling)
     debugImplementation(libs.androidx.compose.ui.test.manifest)
