--- zoomimage-compose-coil-core/src/main/kotlin/com/github/panpf/zoomimage/compose/coil/internal/Utils.kt
@@ -1,108 +1,0 @@
-/*
- * Copyright 2023 Coil Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.github.panpf.zoomimage.compose.coil.internal
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.ReadOnlyComposable
-import androidx.compose.runtime.Stable
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.Constraints
-import androidx.compose.ui.unit.IntSize
-import coil3.compose.AsyncImagePainter
-import coil3.request.ImageRequest
-import coil3.request.NullRequestDataException
-import coil3.size.Scale
-import kotlin.math.roundToInt
-
-
-/** Create an [ImageRequest] from the [model]. */
-@Composable
-@ReadOnlyComposable
-internal fun requestOf(model: Any?): ImageRequest {
-    if (model is ImageRequest) {
-        return model
-    } else {
-        return ImageRequest.Builder(LocalContext.current).data(model).build()
-    }
-}
-
-@Stable
-internal fun transformOf(
-    placeholder: Painter?,
-    error: Painter?,
-    uriEmpty: Painter?,
-): (AsyncImagePainter.State) -> AsyncImagePainter.State {
-    return if (placeholder != null || error != null || uriEmpty != null) {
-        { state ->
-            when (state) {
-                is AsyncImagePainter.State.Loading -> {
-                    if (placeholder != null) state.copy(painter = placeholder) else state
-                }
-
-                is AsyncImagePainter.State.Error -> if (state.result.throwable is NullRequestDataException) {
-                    if (uriEmpty != null) state.copy(painter = uriEmpty) else state
-                } else {
-                    if (error != null) state.copy(painter = error) else state
-                }
-
-                else -> state
-            }
-        }
-    } else {
-        AsyncImagePainter.DefaultTransform
-    }
-}
-
-@Stable
-internal fun onStateOf(
-    onLoading: ((AsyncImagePainter.State.Loading) -> Unit)?,
-    onSuccess: ((AsyncImagePainter.State.Success) -> Unit)?,
-    onError: ((AsyncImagePainter.State.Error) -> Unit)?,
-): ((AsyncImagePainter.State) -> Unit)? {
-    return if (onLoading != null || onSuccess != null || onError != null) {
-        { state ->
-            when (state) {
-                is AsyncImagePainter.State.Loading -> onLoading?.invoke(state)
-                is AsyncImagePainter.State.Success -> onSuccess?.invoke(state)
-                is AsyncImagePainter.State.Error -> onError?.invoke(state)
-                is AsyncImagePainter.State.Empty -> {}
-            }
-        }
-    } else {
-        null
-    }
-}
-
-@Stable
-internal fun ContentScale.toScale() = when (this) {
-    ContentScale.Fit, ContentScale.Inside -> Scale.FIT
-    else -> Scale.FILL
-}
-
-internal fun Constraints.constrainWidth(width: Float) =
-    width.coerceIn(minWidth.toFloat(), maxWidth.toFloat())
-
-internal fun Constraints.constrainHeight(height: Float) =
-    height.coerceIn(minHeight.toFloat(), maxHeight.toFloat())
-
-internal inline fun Float.takeOrElse(block: () -> Float) = if (isFinite()) this else block()
-
-internal fun Size.toIntSize() = IntSize(width.roundToInt(), height.roundToInt())
-
-internal val ZeroConstraints = Constraints.fixed(0, 0)
