+++ zoomimage-core-coil/src/commonMain/kotlin/com/github/panpf/zoomimage/coil/CoilTileBitmap.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.zoomimage.coil
+
+import com.github.panpf.zoomimage.subsampling.CacheTileBitmap
+
+expect class CoilTileBitmap : CacheTileBitmap
