+++ zoomimage-core-coil/src/commonMain/kotlin/com/github/panpf/zoomimage/coil/CoilTileBitmapCache.kt
@@ -1,0 +1,6 @@
+package com.github.panpf.zoomimage.coil
+
+import coil3.ImageLoader
+import com.github.panpf.zoomimage.subsampling.TileBitmapCache
+
+expect class CoilTileBitmapCache(imageLoader: ImageLoader) : TileBitmapCache
